Removed factorize workflow, updated the distribute workflow accordingly
